requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
Faker==33.1.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
